DRTII-1213 Remove comments & printlns
