+++ client/src/test/scala/drt/client/components/PcpPaxSummaryTests.scala
@@ -2,7 +2,6 @@ package drt.client.components
 
 import drt.client.services.JSDateConversions.SDate
 import drt.shared.CrunchApi.{CrunchMinute, MillisSinceEpoch}
-import drt.shared.TQM
 import uk.gov.homeoffice.drt.ports.Queues
 import uk.gov.homeoffice.drt.ports.Queues.Queue
 import uk.gov.homeoffice.drt.ports.Terminals.{T1, Terminal}
@@ -22,12 +21,11 @@ object PcpPaxSummaryTests extends TestSuite {
           generateCrunchMinutes(startMillis, quantity, T1, Queues.NonEeaDesk, 0.5, 2) ++
           generateCrunchMinutes(startMillis, quantity, T1, Queues.EGate, 0.25, 2) ++
           generateCrunchMinutes(startMillis, quantity, T1, Queues.FastTrack, 0.1, 2)
-      val crunchMinutes = crunchMinutesSet.map(cm => (TQM(cm), cm)).toMap
 
-      val summaryStartMillis = SDate("2018-01-01T00:04")
+      val summaryStart = SDate("2018-01-01T00:04")
       val summaryDurationMinutes = 3
 
-      val result = PcpPaxSummary(summaryStartMillis, summaryDurationMinutes, crunchMinutes, T1, Set(Queues.EeaDesk, Queues.NonEeaDesk))
+      val result = PcpPaxSummary(summaryStart, summaryDurationMinutes, crunchMinutesSet, Set(Queues.EeaDesk, Queues.NonEeaDesk))
       val totalPax = (3 * 1 + 3 * 0.5 + 3 * 0.25 + 3 * 0.1).toInt
       val expected = PcpPaxSummary(totalPax, Map(Queues.EeaDesk -> 3, Queues.NonEeaDesk -> 1.5))
 
@@ -49,12 +47,11 @@ object PcpPaxSummaryTests extends TestSuite {
           generateCrunchMinutes(startMillis, quantity, terminal2, Queues.NonEeaDesk, 0.75, 2) ++
           generateCrunchMinutes(startMillis, quantity, terminal2, Queues.EGate, 0.2, 2) ++
           generateCrunchMinutes(startMillis, quantity, terminal2, Queues.FastTrack, 0.05, 2)
-      val crunchMinutes = crunchMinutesSet.map(cm => (TQM(cm), cm)).toMap
 
       val summaryStartMillis = SDate("2018-01-01T00:04")
       val summaryDurationMinutes = 3
 
-      val result = PcpPaxSummary(summaryStartMillis, summaryDurationMinutes, crunchMinutes, T1, Set(Queues.EeaDesk, Queues.NonEeaDesk))
+      val result = PcpPaxSummary(summaryStartMillis, summaryDurationMinutes, crunchMinutesSet, Set(Queues.EeaDesk, Queues.NonEeaDesk))
       val totalPax = (3 * 1 + 3 * 0.5 + 3 * 0.25 + 3 * 0.1).toInt
       val expected = PcpPaxSummary(totalPax, Map(Queues.EeaDesk -> 3, Queues.NonEeaDesk -> 1.5))
