+++ project/Settings.scala
@@ -144,11 +144,9 @@ object Settings {
 
   /** Dependencies only used by the JVM project */
   val jvmDependencies = Def.setting(List(
-    "com.amazonaws" % "aws-java-sdk" % awsSdk,
     "com.github.gphat" %% "censorinus" % censorinus,
     "com.pauldijou" %% "jwt-core" % "4.3.0",
     "com.hierynomus" % "sshj" % sshJ,
-    "com.lihaoyi" %% "pprint" % pprint,
     "com.lihaoyi" %%% "utest" % uTest % Test,
 
     "javax.mail" % "mail" % "1.4.7",
@@ -188,8 +186,8 @@ object Settings {
     "ch.qos.logback.contrib" % "logback-json-classic" % "0.1.5",
     "ch.qos.logback.contrib" % "logback-jackson" % "0.1.5",
     "org.codehaus.janino" % "janino" % janinoVersion,
-    "com.fasterxml.jackson.core" % "jackson-databind" % "2.10.5.1",
-    "com.fasterxml.jackson.core" % "jackson-core" % "2.10.5",
+//    "com.fasterxml.jackson.core" % "jackson-databind" % "2.10.5.1",
+//    "com.fasterxml.jackson.core" % "jackson-core" % "2.10.5",
     "org.pac4j" % "pac4j-saml" % pac4jSaml,
     "org.apache.commons" % "commons-csv" % csvCommons,
     "org.apache.poi" % "poi" % poi,
