+++ server/src/main/twirl/views/index.scala.html
@@ -9,7 +9,6 @@
     <title>@title</title>
     <meta content='width=device-width, initial-scale=1, maximum-scale=1, user-scalable=no' name='viewport'>
     <link rel="stylesheet" media="screen" href=@_asset("stylesheets/main.min.css")?@BuildInfo.version>
-    <link href="//cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css" rel="stylesheet">
     <link rel="shortcut icon" type="image/png" href=@_asset("images/favicon.png")>
     @if(googleTrackingCode.nonEmpty){
     <!-- Google tag (gtag.js) -->
